Refactor API utility functions
Replace arrow functions in 'collections.ts', 'environments.ts', 'folders.ts', and 'methods.ts' with traditional function declarations for increased readability and debuggability. Reorganizes the code and fixes some known scope issues with 'this'. The import paths for IApiResponse and apiService have been aligned to new directory structures. No changes in functionalities.
